rename package login to member
